Hide RSV from test card in queue
